-aggiunto min se locale -correzione rss
git-svn-id: http://subversion.assembla.com/svn/fantamanajer/trunk@318 37f4913f-0a8a-4499-b109-c3fdd495100e

+++ code/backup.code.php
@@ -13,7 +13,8 @@ if($backupObj->dodump())
 	{
 		$handle = fopen(DOCSDIR . 'nomeBackup.txt','r');
 		$fileOld = fgets($handle);
-		unlink(DBDIR . $fileOld);
+		if(file_exists(DBDIR . $fileOld))
+			unlink(DBDIR . $fileOld);
 		fclose($handle);
 		$handle = fopen(DOCSDIR . 'nomeBackup.txt','w');
 		fwrite($handle,$name . '.gz');
